requirements.txt
cachelib==0.9.0
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
Flask-Session==0.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
mongoengine==0.24.2
Pillow==9.3.0
pymongo==4.3.3
python-dotenv==0.21.0
qrcode==7.3.1
six==1.16.0
Werkzeug==2.2.2